muni/muni_site/requirements.txt
pymysql
mysqlclient
pillow
django-crispy-forms